+++ src/components/tools/Stats/StatBlock.svelte
